requirements.txt
python-telegram-bot>=20.0
gspread>=6.0.0
google-auth>=2.0.0
oauth2client>=4.1.3
